# internal/api/http/openapi.yaml
openapi: 3.1.0
info:
  title: Plowered API
  version: "1.0.0"
  description: |
    Plowered is a self-hosted, AI-native data context platform: catalog,
    orchestration, quality, governance, lineage, ETL/ELT, MCP, all in one.

    All endpoints sit behind the same auth chain — session cookie (web UI)
    or Bearer JWT (integrations). Tenant isolation is enforced server-side
    from the authenticated principal; you never pass tenant_id in payloads
    or query strings.

    Identifiers throughout are UUIDs unless documented otherwise.
  contact:
    name: Plowered
    url: https://github.com/Satyaamm/plowered
servers:
  - url: http://localhost:8080
    description: Local dev
security:
  - cookieAuth: []
  - bearerAuth: []
...
paths:
...
  /v1/assets/{id}/terms:
    get:
      tags: [governance]
      summary: Glossary terms linked to an asset
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Term list }
...
  /v1/jobs:
    get:
      tags: [orchestration]
      summary: List recent async jobs for the tenant
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Recent jobs, newest first
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/Job" }
  /v1/jobs/{id}:
    get:
      tags: [orchestration]
      summary: Fetch one async job by ID
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: Job record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Job" }
        "404": { description: Not found }

  # ---------- BYOM: bring-your-own-model provider configs ----------
...
  /v1/runs/{id}/logs:
    get:
      tags: [orchestration]
      summary: Paginated executor logs
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: since, in: query, schema: { type: integer, default: 0 }, description: "Tail cursor — pass the previous batch's max id" }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 1000 } }
      responses: { "200": { description: Lines } }
  /v1/runs/{id}/logs/stream:
    get:
      tags: [orchestration]
      summary: SSE live log tail
      description: |
        Server-Sent Events stream. Each `data:` payload is a JSON log line.
        Emits `event: done` once the run reaches a terminal state.
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: text/event-stream
          content: { text/event-stream: { schema: { type: string } } }

  # ---------- quality ----------
...
  /v1/glossary/terms:
    get:
      tags: [governance]
      summary: List business glossary terms
      responses: { "200": { description: Terms } }
    post:
      tags: [governance]
      summary: Create a term
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/GlossaryTerm" } } }
      responses: { "201": { description: Created } }
  /v1/glossary/terms/{id}:
    get:
      tags: [governance]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Term } }
    patch:
      tags: [governance]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Updated } }
    delete:
      tags: [governance]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "204": { description: Deleted } }
  /v1/glossary/terms/{id}/assignments:
    post:
      tags: [governance]
      summary: Link a term to an asset
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset_id]
              properties:
                asset_id: { type: string, format: uuid }
      responses: { "204": { description: Linked } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: plowered_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        id:           { type: string, format: uuid }
        type:         { type: string, description: "e.g. classify:connection, search:reindex" }
        status:       { type: string, enum: [queued, running, succeeded, failed, cancelled] }
        progress_pct: { type: integer, minimum: 0, maximum: 100 }
        message:      { type: string }
        result:       { type: object, additionalProperties: true }
        error:        { type: string }
        resource_id:  { type: string }
        actor_id:     { type: string }
        created_at:   { type: string, format: date-time }
        started_at:   { type: string, format: date-time, nullable: true }
        finished_at:  { type: string, format: date-time, nullable: true }
...
    GlossaryTerm:
      type: object
      required: [name]
      properties:
        id:          { type: string, format: uuid, readOnly: true }
        name:        { type: string }
        definition:  { type: string }
        parent_id:   { type: string, format: uuid }
        status:      { type: string, enum: [draft, approved, deprecated] }
        owner_id:    { type: string, format: uuid }
        created_at:  { type: string, format: date-time, readOnly: true }
        updated_at:  { type: string, format: date-time, readOnly: true }